add connect wallet functionality and seprate WrongNetworkInnerModal
